requirements.txt
# IBKR API
ib_insync>=0.9.86

# AI APIs
google-generativeai>=0.3.0
anthropic>=0.18.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration & Environment
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Database
aiosqlite>=0.19.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
